// generated/src/app/api-models/stage-auth-sec-opin.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthSecOpin {
  batchId: string;
  transactionId: number;
  seqNo: number;
  provId: string;
  secondOpinion: string;
  secOpnDate: Date;
  decision: string;
  status: string;
  statusReason: string;
  comments: string;
  insertDatetime: Date;
}